SourceReader implementations: fixed naming issues
